requirements.txt
# Add your project dependencies below
requests
python-dotenv
mcp>=0.5.0
aiohttp>=3.8.0
requests>=2.31.0
python-dotenv>=1.0.0
sgp4>=2.22